repair branch and semantic fixes
