+++ .env.example
@@ -6,7 +6,7 @@ TENDERMINT_RPC_URL=https://rpc.celestia-mocha.com/
 CHAIN_ID=11155111
 
 NEUTRON_RPC=https://rpc-falcron.pion-1.ntrn.tech
-ETHEREUM_URL=https://sepolia.infura.io/v3/8a860169610f4d6085a9914fcced4499
+ETHEREUM_URL=https://sepolia.infura.io/v3/YOUR_INFURA_API_KEY
 
 ETHEREUM_DEFAULT_ACCOUNT_ADDRESS="0x51df57D545074bA4b2B04b5f973Efc008A2fde6E"
 NEUTRON_DEFAULT_ACCOUNT_ADDRESS="neutron1m9l358xunhhwds0568za49mzhvuxx9ux8xafx2"
@@ -19,7 +19,7 @@ ETHEREUM_SEPOLIA_MAILBOX_MESSAGES_STORAGE_KEY="0x0000000000000000000000000000000
 ETHEREUM_SEPOLIA_MAILBOX_EXAMPLE_CONTRACT_ADDRESS="0x9151c571b53627Bc20Ce59F18B156AE8AFaADe7d"
 NEUTRON_PION_1_MAILBOX_EXAMPLE_CONTRACT_ADDRESS="neutron1h967w282lz9tv8qxw5ch7a3cefndp4umexhpz9ehqa2d87gg80hsswsrlr"
 
-ETH_PRIVATE_KEY="f7cbe2a6ee6746d96787b7bf9a541a416f9de978519ac7097ec433413beeefc9"
+ETH_PRIVATE_KEY=
 
 SP1_PROVER=cpu
 #NETWORK_PRIVATE_KEY=
