+++ coprocessor/src/examples/rate.rs
@@ -2,9 +2,9 @@ use std::env;
 
 use crate::{
     RATE_APPLICATION_CIRCUIT_ELF,
+    clients::{ClientInterface, DefaultClient},
     coprocessor::{Coprocessor, CoprocessorInterface},
     examples::prove_coprocessor,
-    lightclients::{MockLightClient, MockLightClientInterface},
     read_ethereum_default_account_address, read_neutron_default_account_address,
 };
 use alloy::{
@@ -19,9 +19,9 @@ use sp1_sdk::{HashableKey, ProverClient, SP1Stdin};
 use sp1_verifier::Groth16Verifier;
 use zk_rate_application_types::{RateApplicationCircuitInputs, RateApplicationCircuitOutputs};
 
-pub async fn prove(mock_light_client: MockLightClient) {
+pub async fn prove(mock_light_client: DefaultClient) {
     let (neutron_root, neutron_height) = mock_light_client
-        .neutron_light_client
+        .neutron_client
         .get_latest_root_and_height()
         .await;
     let neutron_vault_balance_key = Ics23Key::new_wasm_account_mapping(
@@ -32,7 +32,7 @@ pub async fn prove(mock_light_client: MockLightClient) {
     let neutron_vault_shares_key =
         Ics23Key::new_wasm_stored_value("shares", &read_neutron_vault_example_contract_address());
     let (ethereum_root, ethereum_height) = mock_light_client
-        .ethereum_light_client
+        .ethereum_client
         .get_latest_root_and_height()
         .await;
     let address =
